// src/app/interfaces/IPanelElements.ts
interface IPanelElements {
  rangeMin?: HTMLInputElement;
  rangeMax?: HTMLInputElement;
  scaleMin?: HTMLInputElement;
  scaleMax?: HTMLInputElement;
  steps?: HTMLInputElement;
  isRange?: HTMLInputElement;
  isVisibleTooltip?: HTMLInputElement;
  isVisibleScale?: HTMLInputElement;
  isReverseDirection?: HTMLInputElement;
  type?: HTMLSelectElement;
}

export default IPanelElements;